+++ inference_perf/client/modelserver/base.py
@@ -12,198 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from abc import ABC, abstractmethod
-from typing import Any, Callable, Dict, Generic, Iterator, List, Optional, Tuple, TypeVar
-from pydantic import BaseModel
+from typing import List, Optional, Tuple
 from inference_perf.config import APIConfig, APIType
 from inference_perf.apis import InferenceAPIData
-
-
-class GaugeResult(BaseModel):
-    avg: float = 0.0
-    median: float = 0.0
-    p90: float = 0.0
-    p99: float = 0.0
-
-    def as_summary(self) -> Dict[str, float]:
-        """Project to the report's mean/median/p90/p99 shape.
-
-        HistogramResult inherits this and is intentionally narrowed to the same
-        four keys (its per_second field is not part of the per-metric summary).
-        """
-        return {"mean": self.avg, "median": self.median, "p90": self.p90, "p99": self.p99}
-
-
-class HistogramResult(GaugeResult):
-    per_second: float = 0.0
-
-
-class RequestsResult(BaseModel):
-    total: float = 0.0
-    per_second: float = 0.0
-
-
-class CounterResult(BaseModel):
-    value: float = 0.0
-
-
-R = TypeVar("R", bound=BaseModel)
-
-
-class Metric(ABC, Generic[R]):
-    metric_name: str
-    target_field: str
-
-    @abstractmethod
-    def get_queries(self, duration: float) -> List[str]:
-        """Returns the ordered list of PromQL queries this metric requires."""
-        ...
-
-    @abstractmethod
-    def parse(self, results: List[float]) -> R:
-        """Convert the ordered query results into a typed result object."""
-        ...
-
-    def collect(self, execute: Callable[[str], float], duration: float) -> R:
-        """Run this metric's queries via execute and parse them into its typed result.
-
-        Keeps query execution and parsing together on the metric so callers never
-        need to know the query/result shape of a particular metric type.
-        """
-        return self.parse([execute(query) for query in self.get_queries(duration)])
-
-
-class GaugeMetric(Metric[GaugeResult]):
-    def __init__(self, target_field: str, metric_name: str, filters: List[str]) -> None:
-        self.target_field = target_field
-        self.metric_name = metric_name
-        self.filters = ",".join(filters)
-
-    def get_queries(self, duration: float) -> List[str]:
-        f, m = self.filters, self.metric_name
-        return [
-            f"avg_over_time({m}{{{f}}}[{duration:.0f}s])",
-            f"quantile_over_time(0.5, {m}{{{f}}}[{duration:.0f}s])",
-            f"quantile_over_time(0.9, {m}{{{f}}}[{duration:.0f}s])",
-            f"quantile_over_time(0.99, {m}{{{f}}}[{duration:.0f}s])",
-        ]
-
-    def parse(self, results: List[float]) -> GaugeResult:
-        return GaugeResult(avg=results[0], median=results[1], p90=results[2], p99=results[3])
-
-
-class CounterMetric(Metric[HistogramResult]):
-    """Returns HistogramResult (not CounterResult) because target fields like prompt_tokens may
-    be populated by either CounterMetric (vllm/sglang) or HistogramMetric (tgi). Counter fills
-    only avg+per_second; percentile fields default to 0."""
-
-    def __init__(self, target_field: str, metric_name: str, filters: List[str]) -> None:
-        self.target_field = target_field
-        self.metric_name = metric_name
-        self.filters = ",".join(filters)
-
-    def get_queries(self, duration: float) -> List[str]:
-        f, m = self.filters, self.metric_name
-        return [
-            f"sum(increase({m}{{{f}}}[{duration:.0f}s]))",
-            f"sum(rate({m}{{{f}}}[{duration:.0f}s]))",
-        ]
-
-    def parse(self, results: List[float]) -> HistogramResult:
-        return HistogramResult(avg=results[0], per_second=results[1])
-
-
-class RequestsMetric(Metric[RequestsResult]):
-    def __init__(self, metric_name: str, filters: List[str], target_field: str = "requests") -> None:
-        self.target_field = target_field
-        self.metric_name = metric_name
-        self.filters = ",".join(filters)
-
-    def _selector(self) -> str:
-        m = self.metric_name
-        if m.startswith("{") and m.endswith("}"):
-            return f"{{{m[1:-1]},{self.filters}}}" if self.filters else m
-        return f"{m}{{{self.filters}}}"
-
-    def get_queries(self, duration: float) -> List[str]:
-        selector = self._selector()
-        return [
-            f"sum(increase({selector}[{duration:.0f}s]))",
-            f"sum(rate({selector}[{duration:.0f}s]))",
-        ]
-
-    def parse(self, results: List[float]) -> RequestsResult:
-        return RequestsResult(total=results[0], per_second=results[1])
-
-
-class HistogramMetric(Metric[HistogramResult]):
-    def __init__(self, target_field: str, metric_name: str, filters: List[str]) -> None:
-        self.target_field = target_field
-        self.metric_name = metric_name
-        self.filters = ",".join(filters)
-
-    def get_queries(self, duration: float) -> List[str]:
-        f, m = self.filters, self.metric_name
-        return [
-            f"sum(rate({m}_sum{{{f}}}[{duration:.0f}s])) / (sum(rate({m}_count{{{f}}}[{duration:.0f}s])) > 0)",
-            f"histogram_quantile(0.5, sum(rate({m}_bucket{{{f}}}[{duration:.0f}s])) by (le))",
-            f"histogram_quantile(0.9, sum(rate({m}_bucket{{{f}}}[{duration:.0f}s])) by (le))",
-            f"histogram_quantile(0.99, sum(rate({m}_bucket{{{f}}}[{duration:.0f}s])) by (le))",
-            f"sum(rate({m}_sum{{{f}}}[{duration:.0f}s]))",
-        ]
-
-    def parse(self, results: List[float]) -> HistogramResult:
-        return HistogramResult(avg=results[0], median=results[1], p90=results[2], p99=results[3], per_second=results[4])
-
-
-class CustomMetric(Metric[CounterResult]):
-    def __init__(self, target_field: str, metric_name: str, op: str, type: str, filters: List[str]) -> None:
-        self.target_field = target_field
-        self.metric_name = metric_name
-        self.op = op
-        self.type = type
-        self.filters = ",".join(filters)
-
-    def _selector(self) -> str:
-        # If the metric name is wrapped in `{...}` (e.g. `{__name__=~"foo(_total)?"}`),
-        # merge the filters into the selector instead of appending a second `{...}` group.
-        m = self.metric_name
-        if m.startswith("{") and m.endswith("}"):
-            return f"{{{m[1:-1]},{self.filters}}}" if self.filters else m
-        return f"{m}{{{self.filters}}}"
-
-    def get_queries(self, duration: float) -> List[str]:
-        if self.type != "counter":
-            return []
-        selector = self._selector()
-        if self.op == "rate":
-            return [f"sum(rate({selector}[{duration:.0f}s]))"]
-        if self.op == "increase":
-            return [f"sum(increase({selector}[{duration:.0f}s]))"]
-        return []
-
-    def parse(self, results: List[float]) -> CounterResult:
-        return CounterResult(value=results[0]) if results else CounterResult()
-
-
-class BaseMetrics:
-    def __init__(self, custom_metrics: Optional[List[Metric[Any]]] = None) -> None:
-        self.custom_metrics = custom_metrics or []
-
-    def _iter_metrics(self) -> Iterator[Metric[Any]]:
-        """Yield every metric this container holds.
-
-        Subclasses override this to yield their named fields (and then defer to
-        super() for the custom metrics) so iteration is explicit rather than a
-        reflection walk over __dict__.
-        """
-        yield from self.custom_metrics
-
-    def __iter__(self) -> Iterator[Metric[Any]]:
-        seen: set[int] = set()
-        for metric in self._iter_metrics():
-            if id(metric) not in seen:
-                seen.add(id(metric))
-                yield metric
+from .metrics import BaseMetrics
 
 
 class ModelServerClient(ABC):
